Gangplank: embed cluster configuration into context
To support `cosa remote` sending payloads via stdin (the method that
OpenShift Builds uses) and for Podman, we need a method to define
std{in,out,err{ and where they go.

This change adds a new ClusterContext that defines the "how" a pod is
created. To support redirecting `stdin` to the pod, the podman creation
code is uses libpod.runtime instead of executing a `podman attach`.

Finally, this prepares the ability to run Gangplank against an arbitrary
cluster for cross-cluster building.

Signed-off-by: Ben Howard <ben.howard@redhat.com>