buildextend-live: fix crash when running on !x86_64
E.g. on PowerPC, we delete a bunch of files from the ISO which we don't
need. This throws off the dictionary mapping we previously built which
assumes all the files with karg embed areas still exist.

Add a filtering step to drop the files which no longer exist before
writing out the header in the System Area.